Renamed package to avoid using dashes on name
According to this discussion on Stackoverflow (https://stackoverflow.com/questions/761519/is-it-ok-to-use-dashes-in-python-files-when-trying-to-import-them#761531) and PEP-8 (https://www.python.org/dev/peps/pep-0008/#id40), the use of dashes should be avoided. ("Modules should have short, all-lowercase names. Underscores can be used in the module name if it improves readability. Python packages should also have short, all-lowercase names, although the use of underscores is discouraged.")